app/finnhub_producer/requirements.txt
websocket-client 
kafka-python 
finnhub-python
avro-python3 
python-dotenv